+++ BookService/src/main/java/com/example/service/impl/BookServiceImpl.java
@@ -1,6 +1,6 @@
 package com.example.service.impl;
 
-import java.math.BigInteger;
+import java.math.BigDecimal;
 import java.time.Instant;
 import java.util.List;
 
@@ -66,7 +66,7 @@ public class BookServiceImpl implements BookService {
         if (!StringUtil.isEmpty(bookInfoDao.shortDescription)) {
             bookInfo.setShortDescription(bookInfoDao.shortDescription);
         }
-        if (bookInfoDao.price != null && bookInfoDao.price.compareTo(BigInteger.ZERO) > 0) {
+        if (bookInfoDao.price != null && bookInfoDao.price.compareTo(BigDecimal.ZERO) > 0) {
             bookInfo.setPrice(bookInfoDao.price);
         }
         bookInfoRepository.save(bookInfo);
